change bf16 to fp16 in optim

+++ src/optim/base.py
@@ -4,7 +4,7 @@ from data.utils import get_dataloader
 import torch
 import torch.nn.functional as F
 import wandb
-import time 
+import time
 import itertools
 import copy
 import random
@@ -16,8 +16,8 @@ from .utils import eval, get_batch, save_checkpoint
 def train_base(model, opt, data, data_seed, scheduler, iterations, acc_steps, batch_size, sequence_length, eval_freq, ckpt_path, distributed_backend,extra_args, itr=0,rng_state_dict=None):
     device_type = 'cuda' if 'cuda' in str(extra_args.device) else 'cpu'
     type_ctx = nullcontext() if device_type == 'cpu' else torch.amp.autocast(
-        device_type=device_type, dtype=torch.bfloat16)  # extra_args.dtype)
-    best_val_loss, text_table = float('inf'), None # best_val_loss not used atm, early stopping not recommended but possible 
+        device_type=device_type, dtype=torch.float16)  # extra_args.dtype)
+    best_val_loss, text_table = float('inf'), None # best_val_loss not used atm, early stopping not recommended but possible
     substep = itr * acc_steps
     data["train"], train_sampler = get_dataloader(
         data["train"],
@@ -26,7 +26,7 @@ def train_base(model, opt, data, data_seed, scheduler, iterations, acc_steps, ba
         seed=data_seed,
         distributed_backend=distributed_backend,
     )
-    
+
     data["val"], val_sampler = get_dataloader(
         data["val"],
         sequence_length=sequence_length,
@@ -36,7 +36,7 @@ def train_base(model, opt, data, data_seed, scheduler, iterations, acc_steps, ba
 
     num_substeps_per_epoch = len(data["train"])
     train_epochs = substep//num_substeps_per_epoch
-    
+
     if rng_state_dict is not None and  rng_state_dict.get("train_sampler_state", None) is not None:
         train_sampler.generator.set_state(rng_state_dict["train_sampler_state"])
     if hasattr(train_sampler, "set_epoch"):
@@ -45,14 +45,14 @@ def train_base(model, opt, data, data_seed, scheduler, iterations, acc_steps, ba
         sampler_state_before_iter = train_sampler.generator.get_state()
     data_train_iter = iter(data["train"])
 
-    
+
     # for val data we don't care about epochs? just cycle through (no need to set_epoch to reshuffle)
     data_val_iter = itertools.cycle(data["val"])
 
     stats = {"train_loss": [], "val_loss": [], "val_pp": [], "val_acc": []}
 
-   
-    
+
+
     if extra_args.compile:
         print(f"Compiling model ...")
         model = torch.compile(model) # requires pytorch 2.0+
@@ -60,7 +60,7 @@ def train_base(model, opt, data, data_seed, scheduler, iterations, acc_steps, ba
     model.train()
 
     t0 = time.time()
-    
+
     if rng_state_dict is not  None:
         torch.set_rng_state(rng_state_dict["cpu_rng_state"])
         torch.cuda.set_rng_state(rng_state_dict["gpu_rng_state"])
@@ -69,12 +69,12 @@ def train_base(model, opt, data, data_seed, scheduler, iterations, acc_steps, ba
     for _ in range(substep % num_substeps_per_epoch):
         get_batch(data_train_iter, device=extra_args.device)
 
-    
+
     while itr < iterations:
-            
+
         for microstep_idx in range(acc_steps):  # gradient accumulation
             x, y = get_batch(data_train_iter, device=extra_args.device)
-            
+
             with type_ctx:
                 with distributed_backend.get_context_for_microstep_forward(model=model, microstep_idx=microstep_idx, gradient_accumulation_steps=acc_steps):
                     outputs = model(x, targets=y)
@@ -110,7 +110,7 @@ def train_base(model, opt, data, data_seed, scheduler, iterations, acc_steps, ba
                 model.eval()
                 train_loss = loss.detach().cpu().item() * acc_steps
                 current_lr = scheduler.get_last_lr()[0] if scheduler is not None else extra_args.lr
-                
+
                 eval_steps = (
                     24 if itr < iterations else len(data["val"])
                 )
@@ -171,7 +171,7 @@ def train_base(model, opt, data, data_seed, scheduler, iterations, acc_steps, ba
                                 py_rng_state=random.getstate(),
                                 train_sampler_state=sampler_state_before_iter,
                                 ckpt_path=os.path.join(os.path.dirname(ckpt_path), f"ckpt_{itr}.pt"))
-                
+
     if distributed_backend.is_master_process():
         print(f"saving checkpoint to {ckpt_path}")
         save_checkpoint(distributed_backend=distributed_backend,
